requirements.txt
discord-py-interactions>=5.13.2
google-generativeai